Features: Adding and testing LLM for further project idea
