+++ requirements.txt
@@ -3,3 +3,4 @@ python-dotenv==1.0.1
 pandas==2.2.3
 streamlit==1.39.0
 plotly==5.24.1
+transformers==4.44.2
